chore: fix confusing comment about FD number 4 that does not exist (#4157)
This was leftover from a time when we tried using non-duplex streams, but since we are using a single duplex stream there's only 4 FDs (0 through 3), and not 5.

Thanks to @TimothyJones for finding it.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0