chore: upgrade dependencies in superchain image
The Debian 10 image uses Python 3.7 which was causing failures elsewhere due to required minimum
python 3.8. Debian 10 is also ten months out of security support so upgrading it seemed like the
right decision here. This change also updates Node to 16 due to Node 14 being EOL and updates
Maven to 3.9.3 dur to the distro link for the older version not working when building a local
superchain image.

Signed-off-by: Kendra Neil <53584728+TheRealAmazonKendra@users.noreply.github.com>