fix(kernel): incorrectly scoped FQN resolutions
The registration was done on the object's prototype, and tthe value from
the constructor was always used, even if that was inherited, such that
if a parent type ahd already been resolved previously, all child types
would use the parent's FQN instead of their own.

Addressed this by instead storing the associations in an external
WeakMap, and added a test case to validate correct behavior.

Fixes aws/aws-cdk#26604