chore: cut down on Docker Images built, drop Node 14 (#4216)
Our Docker image building matrix was gigantic, which was causing long build times and build nodes to run out of disk space and memory.

Cut down on the flavors we build.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0