feat(jsii-reflect): TypeSystem can be locked to improve reflection performance (#4318)
Memoizes additional calls that rely on the typesystem to retrieve type instances.
These calls can (theoretically) change when the typesystem is changed.
Therefore we cannot assume it's okay to always memoize the first call.

To workaround this limitation, we introduce a new mechanism to manually `lock` the typesystem once all assemblies are loaded.
We then can start memoizing the additional calls.

For example in `awslint` this reduces the runtime against `aws-cdk-lib` by ~20s.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0