feat: switch to python 3.8 (#4239) (#4327)
Python 3.7 has been end-of-life since 2023-06-27, and an increasing amount of our dependencies are dropping support for it as a consequence.

We introduced a warning to users that are still using Python 3.7 with release v1.86.0 around August 1, 2023.

We are now effectively dropping support for Python 3.7.

Related: https://github.com/aws/jsii/commit/0ef7acef765690f99075888d6c3e6f03d7cc2d59

Re-roll of #4239, which was reverted because the build & release pipeline failed. After some investigation this was because the job used a superchain image with python3.7. However after this change a minimum version of python 3.8 is required.

BREAKING CHANGE: jsii-pacmak and the libraries it generates for Python targets now require a minimum Python version of 3.8, instead of 3.7 previously. We recommend users upgrade to the latest supported Python release (Python 3.11).

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0