chore: update DockerHub README daily (#4356)
Add a GitHub action to update the DockerHub README of the superchain image every day.

Needed to to refactor the steps a little bit to get the credentials out in a way the Action needs them.

(Of course, no way to test this except in production :/ )


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0