fix(pacmak): invokeBinScript fails when using symlinked cache
This was previously attempted to fix in https://github.com/aws/jsii/pull/4324
While the above fix resolves issues with dependencies, it causes failures when the binary is shelling out to other node processes. This is due to the intrusive and indiscriminate overloading of NODE_OPTIONS, which will forcibly apply to any child processes as well. While in theory adding the symlink flags should not be an issue, this seems to trigger a bug in node: https://github.com/nodejs/node/issues/41000

tl;dr this all sucks very much and we are now just disabling the runtime cache for binaries.