fix(jsii-pacmak): show custom pack-command in timer label (#4392)
Currently 'npm pack' is shown in the logs regardless of the custom command.  Like in:
```log
[jsii-pacmak] [INFO] Packaged. npm pack (42.4s) | cleanup (3.8s) | load jsii (0.0s) | js (0.0s)
```

This behaviour is confusing because it leaves doubt about which command was actually applied. This PR passes the value of the pack-command option to the timer label.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0