chore(superchain): explicitly list supported images (#4500)
There are a lot of image tags in the jsii/superchain image repo. Only a few of which are actively supported.

Previously we attempted to explain the rules and list supported versions of the subcomponents, but that was complicated and missed a few cases.

With this change we are just listing them explicitly.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0