chore: fix typo in error message - `delcaration` (#4514)
Observed the error below: `The implemented delcaration is here.`

`delcaration` - I kept looking at this because it hurt my eyes until I finally realized it was a typo.

At the very least, this can be closed as won't fix so that at least there will be a record for the next person that searches for an issue or fix for it.

## Observed Error Message

```
[2024-05-15T12:22:12.636] [ERROR] jsii/compiler - Type model errors prevented the JSII assembly from being created
src/myconstruct.ts:191:31 - error JSII5004: "cdkConstruct.MyConstruct#dynamoDBTable" changes the property type to "aws-cdk-lib.aws_dynamodb.Table" when implementing cdkConstruct.IMyConstruct. Change it to "aws-cdk-lib.aws_dynamodb.ITable"

191   public get dynamoDBTable(): dynamodb.Table | undefined {
                                  ~~~~~~~~~~~~~~~~~~~~~~~~~~

  src/myconstruct.ts:174:28
    174   readonly dynamoDBTable?: dynamodb.ITable;
                                   ~~~~~~~~~~~~~~~
    The implemented delcaration is here.
👾 Task "build » compile" failed when executing "jsii --silence-warnings=reserved-word" (cwd: packages/cdk)
```


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0