fix(jsii-config): update jsii-config and docs with latest settings (#4515)
Updates jsii-config and the website with the latest settings for user-provided TypeScript configs.

The update to jsii-config includes some other changes that have previously been missed, including support for go and fixing a the outdir configuration.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0