fix(rosetta): python identifiers mishandle identifier translations with capital letters (#4644)
Fixes https://github.com/aws/aws-cdk/issues/29138

Our python translation currently translates `sizeInMBs` to `size_in_mbs` for the python examples. However, the actual property is named `size_in_m_bs`. This is because we are currently handling identifiers with the regex `[^A-Z][A-Z]`, which misses the `MB` pair. 

I cannot see a use case where we would want to handle _identifiers_ this way (as opposed to interface names ala `IInterface`), so I've changed the logic.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0