chore(mergify): move deprecated parameters from `pull_request_rules` to `queue_rules` (#4670)
We are using deprecated parameters that are about to be removed (October 21st, 2024).

Ideally we should move away from Mergify and use GitHub merge queues. But the config for this repository is more complex, with many pull request rules. Let's avoid breaking the PRs for now and then we can come back and migrate to merge queues.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0