feat(pacmak): support jsii-rosetta 5.7 (#4718)
Adds support for jsii-rosetta 5.7 to jsii-pacmak.

This also changes the peer dependency to simply state the minimal required version.
This will make it less urgent for us to actually release a new version of jsii-pacmak.

Also includes a new script to fully automate the uprade.
A future extension would be to run this script from a workflow.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0