chore: fix the Python build for Python >= 3.12 (#4827)
Since the removal of a class in Python 3.12, we need a more modern version of `setuptools` to work on newer Python versions.

This gets rid of the error:

```
AttributeError: module 'pkgutil' has no attribute 'ImpImporter'. Did you mean: 'zipimporter'?
```


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0