fix(jsii-pacmak): CS0108 <static property> hides inherited  member (#4826)
C# requires that we explicitly declare when static members are changed in subclasses. In #2373 we added the ability to change the definition of static functions and static read/write properties in this way, but we missed doing the same for static `const` properties.

Give them the same treatment.

(PR includes https://github.com/aws/jsii/pull/4827)

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0