fix(jsii-pacmak): incorrect casing for class namespaces in other module
When resolving the .NET FQN for a type nested in a class from another
module:

```
Module.Namespace.SomethingFANCY.SomeType
                    ^^^^^
                      what is this?
```

The C# code generator incorrectly assumes that all parts leading up to
the final type name are module names.

Since module names are PascalCased, this means the name gets changed to
`SomethingFancy`. However, in JSII classes can also hold types, but
class names are not PascalCased (instead, they are capitalized).

The upshot was that for L1 property types, which live inside L1 classes,
the name of the class would be incorrectly rendered.

This would only happen for references to types in other modules, because
when processing inside the same module, the code generator was holding
a lookup table that made it correctly realize that `SomethingFANCY` in
this case is a class, making it take a different code path.

In this change, try to distinguish between indicated components
representing submodules and representing types, and mangle the names
differently.

Suffice to say this is all a bit of a mess, but barring a full redesign
of how we trace generated this seems to be the best we can do for now.

Fixes https://github.com/aws/jsii/issues/4849.