chore: allow current Node version to be EOL (#4853)
Node 18 is EOL, but CDK is still supporting it.

Since CDK is still supporting it, we can't have jsii not support it.

But the jsii build will fail when running on Node 18, which is not really acceptable.


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0