feat: prepare support for intersection types
This is a preparatory PR for adding support for intersection types (`A &
B`) to jsii.

That feature might land or might not, but to figure it out we first need
to update `@jsii/spec` to allow modeling intersection types to begin
with, as other jsii packages live outside this repo so we can't do
an atomic commit adding the feature everywhere; we have to release
an updated version of the spec first.

To this end, we're doing the following:

- Add the necessary fields to the spec to express the new feature.
- Add a `usedFeatures` field to assemblies to indicate whether or
  not a "new feature" is being used, so that tools can reject
  an assembly if it's using a feature they don't support.
- Wherever we call Rosetta directly from `jsii-pacmak`, we have to
  do runtime type assertions to make sure the types conversion
  is valid.

The `usedFeatures` field is chosen rather than a schema version field,
to maximize compatibility: we want to gate loading an assembly on
whether or not the assembly actually *uses* the new feature, not on
whether or not the producing tool *knows about* the new feature.  JSON
Schema validation would do the same thing (albeit with a worse error
message), but for example the kernel doesn't do JSON Schema
validation because it is a slow process.