chore: fix build problems (#4899)
The build in the main pipeline fails. I don't quite understand the reason, but it has to do with the `align-version` script:

* After the versions have been "aligned", all package versions in this repo have a dependency on `1.113.0` (which is an actually existing package)
* All of a sudden, the TypeScript compilation of `jsii-pacmak` starts picking up the types from the `1.113.0` version of `@jsii/spec`, rather than the local workspace version. 
  * This might sound logical, but mind you that no files on disk have changed and no symlinks have been created. The Node.js place that `@jsii/spec` resolves to is still the local working copy, but TypeScript picks another version, from somewhere, for some reason.

Solve this by bumping to an unreleased version before doing the pipeline build. That way, the bumped version number can never match an actual published version.

Also change all caret dependencies `^` on `@jsii/spec` into point version dependencies. We don't accidentally want a newer version substituted in, because that has implications for the `usedFeatures` that a consumer claims to support.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0