feat(dotnet-runtime): prepare for intersection types (#4919)
The .NET runtime works by getting its JSII metadata like the function to call and the arguments from an attribute. 

## What is all of this about in the first place?

So the call looks like this:

```cs

        [JsiiMethod(name: "someMethod", returnsJson: "{\\"type\\":{\\"primitive\\":\\"string\\"}}")]
        public virtual string SomeMethod()
        {
            return InvokeInstanceMethod<string>(new System.Type[]{}, new object[]{})!;
        }
```

Think a moment about what `InvokeInstanceMethod` has to do:

```cs
        [JsiiMethod(name: "someMethod", returnsJson: "{\\"type\\":{\\"primitive\\":\\"string\\"}}")] // <-- needs this
        public virtual string SomeMethod()
        {
            // InvokeInstanceMethod receives its callers name (so "SomeMethod") via compiler magic,
            // and then proceeds to look up the method named "SomeMethod" on "this" in order to find 
            // the JsiiMethod attribute.
            return InvokeInstanceMethod<string>(new System.Type[]{}, new object[]{})!;
                                                // ^^^^^^^^ also passes the types of the current object's arguments
                                                //          because it uses the reflection call that can resolve
                                                //          overloads.
        }
```



Here's what we could have done:

```cs
        // Look ma, no attribute!
        public virtual string SomeMethod()
        {
            return InvokeInstanceMethod<string>("someMethod", "{\\"type\\":{\\"primitive\\":\\"string\\"}}", new object[] { /* input arguments here */ });
        }
```

I don't quite understand why we need to do the reflection to get to the attribute in the first place, if we have that information Right There.

In any case, no use crying about it now, I'm not changing all of that.

## What is all of this about in the first place?

`InvokeInstanceMethod` calls the reflection API [GetMethod](https://learn.microsoft.com/en-us/dotnet/api/system.type.getmethod?view=net-9.0), which can be called in a variety of ways. In the simplest way we give it just a method name, or a method name and some binding flags to distinguish between static and member methods, etc.

This method can also do overload resolution, so if you have a `Method(string)` and a `Method(int)` you can pass the appropriate types to get it to resolve to the method you want.

This raises a problem for us, because we are now trying to access generic methods so we must come up with a `Type` object to represent a generic type (i.e., not a real type). I don't know how to do that and I couldn't figure it out.

But it occurs to me that we don't need to do overload resolution to begin with, because jsii guarantees that there is always exactly one method with a given name: TypeScript doesn't have overloading; so we'll never generate multiple methods with the same name but different argument types; so we don't need to have types to resolve overloads either.

So we drop the `Type[]` array we get passed on the floor, and just resolve the method by name.

## Also in this PR

Also update the classes that are used to deserialize the JSON in our attribute to recognize the `intersection` field. Again, we could have constructed those objects Right There and have avoided the overload of JSON deserialization on every call. But this is what we're doing now.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0