feat(jsii-pacmak): document union types in JavaDoc
In some places union types are rendered to overloaded setters, and in
other places the type is collapsed to `Object`. Whenever the type is
collapsed to `Object`, any other type information is lost for the user.

In this PR, emit the union type information to JavaDocs so users can
find it again.