fix(pacmak): python classes are generated out of order (#4927)
Fixes #4426

This issue only occurred in submodule using a `pascalCased` name. The reason was that the dependency resolution did not the the same snake_case transformation of the submodule name: When the bases and interfaces of such a type where looked up, an untransformed fqn was used. Consequently the base or interface wasn't found and the type was considered dependency-less. This ultimately led to types being (incorrectly) emitted alphabetically.

The fix was to update the `TypeResolver` to use the same `toPythonFqn()` implementation as elsewhere in the code.


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0