feat(jsii-pacmak): document union types for C#
In some places union types are rendered to overloaded setters, and in
other places the type is collapsed to `Object`. Whenever the type is
collapsed to `Object`, any other type information is lost for the user.

In this PR, emit the union type information to XML docs so users can find it again.

Also update the JavaDoc handler to handle type intersections.