fix: type intersections don't work for Java class builders (#4945)
We have two different code paths for struct builders and class builders
(where the class constructor takes a props object as an argument).

We were supporting type intersections properly in struct builders, but
not yet in class builders.

Fix that.


---

By submitting this pull request, I confirm that my contribution is made
under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0