fix: covariant method overrides break Go compilation (#4967)
Currently generated Go code is not compiled, which leads to potential codegen mistakes slipping in.

Switch on compilation by default; i.e., make the opt-in environment variable an opt-out environment variable.

This revealed a problem with covariant return types in Go; this concept does not exist. We'll solve that by not emitting the more specific return type. This means Go users will not benefit from the additional typing, but the feature can remain in other languages and the compilation will succeed. Go users instead get documentation telling them the real type. They can easily user interface conversions to convert to the actual type. While this is a lot awkward, it still makes it possible for Go users to benefit from this improvement in most circumstances.

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0