fix(jsii-pacmak): Python modules fail due to bidirectional type imports
In the situation of 2 modules that use each other's types, all languages
are able to cope except Python. For Python, `jsii-pacmak` generates
reciprocal `import` statements (at the top) which lead to a (failing)
import from an incomplete module.

To solve this, we render all type signatures strictly in quotes, and
gate all imports that are only being done for type declarations
behind an

```py
if typing.TYPE_CHECKING:
   ...
```

gate.