fix(jsii-pacmak): Python interfaces sometimes violate MRO
The Method Resolution Order (MRO) rules for Python state (in brief)
that all classes and interfaces in an inheritance declaration must
be listed subclass-before-superclass, and in the same order on all
classes.

`jsii-pacmak` was not guaranteeing order, leading to sometimes producing
an order that violates MRO.

Update `jsii-pacmak` to update base classes and sort them in the right
order.