fix(jsii-pacmak): Python interfaces sometimes violate MRO (#4973)
The Method Resolution Order (MRO) rules for Python state (in brief) that all classes and interfaces in an inheritance declaration must be listed subclass-before-superclass, and in the same order on all classes.

`jsii-pacmak` was not guaranteeing order, leading to sometimes producing an order that violates MRO.

Update `jsii-pacmak` to update base classes and sort them in the right order.



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0