fix: Typeguard cannot check the protocol (#4986)
Versions 3.0 and higher of `typeguard` can't check the jsii protocol interfaces and produce a noisy warning.

Because `constructs` hadn't been updated for a while, and it had been generated with an old version of `jsii-pacmak` that was forcing `typeguard` to `2.13`, no-one had noticed all the changes to `pacmak` that try to support a newer version of `typeguard`, because in practice it never ended up being installed (all CDK projects will install `constructs` which was keeping `typeguard` at an old version).

Revert to that old version, and stick at it until we can fix up our generated Python bindings to be more modern.

Closes https://github.com/aws/constructs/issues/2825


---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0