fix(jsii-pacmak): compilation fails if a type is named "Object"
When jsii is used to emit a type name `"Object"`, the definition of the
`equals()` method that `jsii-pacmak` generates changes meaning:

```java
// This is meant to reference the built-in Object class
@Override
public final boolean equals(final Object o) {
}

// But with a local "Object" class present, this now references
// the local class.
```

We then get the errors:

```
error: method does not override or implement a method from a supertype
error: incomparable types: Jsii$Proxy and Object
...
```

And a bunch more.

Instead, emit the `equals()` method with a fully qualified reference
to `java.lang.Object` so the aliasing can't happen.